Index

Zalgo.Zalgo — Module

Zalgo.jl does two things.

It adds pointless diacritics to text: zalgo("Cthulhu")

It converts an input ASCII string to equivalent characters found in the darkest recesses of the Unicode charts:

blackboard("Hello World")    
boldfraktur("Hello World")   
bolditalic("Hello World")    
bolditalicsans("Hello World")
boldroman("Hello World")     
boldsans("Hello World")      
boldscript("Hello World")    
boxed("hello world")         
circled("HELLO WORLD")       
fraktur("Hello World")       
italic("Hello World")        
italicsans("Hello World")    
sans("Hello World")          
script("Hello World")        
segmented("0123456789")      
teletype("Hello World")      
upsidedown("Hello World")    

large_type("Hello World") displays the text using the Large Type glyphs added to Unicode in version 16.

Zalgo.LargeTypeChar — Type

A single "LargeType" glyph, using the system of pieces defined in Unicode version 16. Each glyph is made up of nine pieces, each in the range U1CE1A to U1CE50.

Zalgo.LargeTypeString — Type

A sequence of LargeType glyphs

Zalgo._buildLargeTypeChar — Method

Construct a LargeType glyph from a string.

Zalgo.blackboard — Method

blackboard(str)

Return a version of string str with blackboard (double-struck) letters from the Unicode table.

Zalgo.boldfraktur — Method

boldfraktur(str)

Return a version of string str with boldfraktur letters from the Unicode table.

Zalgo.bolditalic — Method

bolditalic(str)

Return a version of string str with bolditalic letters from the Unicode table.

Zalgo.bolditalicsans — Method

bolditalicsans(str)

Return a version of string str with bold italic sans serif letters from the Unicode table.

Zalgo.boldroman — Method

boldroman(str)

Return a version of string str with boldroman letters from the Unicode table.

Zalgo.boldsans — Method

boldsans(str)

Return a version of string str with bold sans serif letters from the Unicode table.

Zalgo.boldscript — Method

boldscript(str)

Return a version of string str with boldscript letters from the Unicode table.

Zalgo.boxed — Method

boxed(str)

Return a version of string str with boxed letters from the Unicode table.

boxed("A") -> "🄰"
boxed("a") -> "🅰"

Zalgo.circled — Method

circled(str)

Return a version of string str with circled/boxed letters from the Unicode table.

A-Z           "A" -> "Ⓐ"  Ⓐ:Ⓩ       
a-z           "a" -> "ⓐ"  ⓐ:ⓩ       
0-9           "0" -> "⓪"  ①:⑨       
A-Z inverse   "A" -> "🅐"  🅐:🅩     
a-z inverse   "a" -> "🅰"  🅰:🆉     

Zalgo.fraktur — Method

fraktur(str)

Return a version of string str with Fraktur letters from the Unicode table.

Zalgo.italic — Method

italic(str)

Return a version of string str with italic letters from the Unicode table.

Zalgo.italicsans — Method

italicsans(str)

Return a version of string str with italic sans serif letters from the Unicode table.

Zalgo.large_type — Method

large_type(str)

Return a string of Unicode glyphs that will display the str as "Large Type". Each glyph is made up to 9 small pieces.

Like the other functions, the current font should have these characters (at Unicode points range U+1CE1A to U+1CE50) otherwise you probably won't sequence anything.

Zalgo.sans — Method

sans(str)

Return a version of string str with sans serif letters from the Unicode table.

Zalgo.script — Method

script(str; roundhand=false)

Return a version of string str with mathematical script letters from the Unicode table.

There are two basic styles of mathematical script lettering: the “regular” calligraphic or Chancery alphabet, and the “fancy script” or round hand alphabet.

By default, the script style will be “script”. If roundhand is true, the style will be “roundhand”.

For more details, see this Unicode document.

Zalgo.segmented — Method

segmented(str)

Return a version of string str with LED-style digits from the Unicode table.

Zalgo.teletype — Method

teletype(str)

Return a version of string str with teletype (monospaced) letters from the Unicode table.

Zalgo.upsidedown — Method

upsidedown(str)

Return a version of string str with upside down letters from the Unicode table.

Zalgo.zalgo — Method

zalgo(str::String, upmarks = 1:4, middlemarks = 1:4,
    downmarks = 1:4, maxmarks = 6)

Randomly add up to maxmarks diacritic marks to each letter of str. The upmarks, middlemarks, and downmarks ranges determine the minimum and maximum number of diacritic marks added to the letter at that position.